% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_AMICo1.R
\name{AMICo1}
\alias{AMICo1}
\title{Title}
\usage{
AMICo1(
  experiment,
  signal,
  lagSec = 4,
  sgol_p = 2,
  sgol_n = 25,
  weightMalus = 30,
  minSizeSec = 5,
  match_threshold = 0,
  outputName,
  correctionRangeSeconds,
  minPeakAmplitude,
  algorithm = c("v1.1", "v1.0")
)
}
\arguments{
\item{experiment}{}

\item{signal}{}

\item{lagSec}{}

\item{sgol_p}{}

\item{sgol_n}{}

\item{weightMalus}{}

\item{minSizeSec}{}

\item{match_threshold}{}

\item{outputName}{}

\item{correctionRangeSeconds}{}

\item{minPeakAmplitude}{}

\item{algorithm}{}
}
\description{
Title
}
