% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{at}
\alias{at}
\title{Get rats value at given time}
\usage{
at(x, time)
}
\arguments{
\item{x}{a rats object}

\item{time}{a time in a format accepted by \link[DyadSync]{timeMaster}}
}
\value{
a rats object of length 1
}
\description{
Get rats value at given time
}
\examples{
at(rats(500:100, frequency=10), time = 12.378)
}
