% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{at_s}
\alias{at_s}
\title{Get rats sample number at given time}
\usage{
at_s(x, time)
}
\arguments{
\item{x}{a rats object}

\item{time}{a time in a format accepted by \link[DyadSync]{timeMaster}}
}
\value{
the index of the vector x where x$x matches time
}
\description{
Get rats sample number at given time
}
\details{
this can be useful e.g. after windowing
}
\examples{
at_s(rats(500:100, frequency=10), time = 12.378)
}
