% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{byWin}
\alias{byWin}
\alias{wapply}
\title{Applies a function over moving windows}
\usage{
byWin(
  x,
  WIN,
  INC,
  flex = TRUE,
  FUN,
  ...,
  position = c("mid", "first", "last"),
  cores = "max"
)

wapply(
  x,
  WIN,
  INC,
  flex = TRUE,
  FUN,
  ...,
  position = c("mid", "first", "last"),
  cores = "max"
)
}
\arguments{
\item{x}{a rats object or another object for which as.rats methods are available}

\item{WIN}{integer. the window size in the time unit of x (e.g. seconds)}

\item{INC}{integer. the increment size in the time unit of x (e.g. seconds)}

\item{flex}{logical. Creates additional shorter windows at the start and end
of the signal to maximize the coverage of the original series}

\item{FUN}{The function to be executed}

\item{...}{Additional parameters to be passed to FUN}

\item{position}{Should the time value of each calculation be positioned as the
first, middle or last sample of each window?}

\item{cores}{integer or "max". The number of cores to be used for parallelization.
if set to 1, parallelization is turned off. Very simple tasks can be faster if run
as single core due to the overhead of scheduling the task and returning the
result can be greater than the time to execute the task itself, resulting in
poor performance.}
}
\description{
Applies a function over moving windows
}
