% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{c.rats}
\alias{c.rats}
\title{Combine rats}
\usage{
\method{c}{rats}(...)
}
\arguments{
\item{...}{}
}
\value{
a rats object.
}
\description{
Various rats objects can be combined if their x values don't overlap, and
if they share the same frequency and timeUnit.
}
\examples{

a1 = rats(runif(10),start=20, frequency=10)
a2 = rats(letters[1:22], start=0, end=2.1, frequency = 10)
a3 = rats(rep(0,13),start = 21, frequency = 10, timeUnit = "second")

l = list(a1,a2,a3)
print(l)
}
