% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_permutations.R
\name{categoryPerm}
\alias{categoryPerm}
\title{Permutation tests for signals in specific epochs}
\usage{
categoryPerm(
  x,
  plotPrefix,
  signal,
  sync,
  stream,
  category,
  categoryIndex,
  keepOnly = c(),
  minEpochSec,
  minOccurrences,
  absolute = FALSE,
  nIter = 1000,
  rotate = TRUE,
  extract_from = c("all", "remaining")[2],
  summarizingFunction = c("mean", "median")[2],
  parameterFunction = c("mean", "median")[2],
  xlim = "auto"
)
}
\arguments{
\item{x}{a DyadExperiment Object.}

\item{plotPrefix}{character. path and prefix of output plots names.}

\item{signal}{character. name of a DyadSignal contained in x.}

\item{sync}{character. Name of a synchronization object. E.g. PMBest, or CCFBest}

\item{stream}{character. Name of a rats object contained within sync}

\item{category}{character. Name of a DyadCategory object contained in x}

\item{categoryIndex}{character. Name of a factor column in 'category'.}

\item{keepOnly}{vector of characters denoting if only given categories must be analyzed. Eg: c("cat1","cat2").}

\item{minEpochSec}{number of seconds of minimum epoch. Shorter epochs will be deleted.}

\item{minOccurrences}{min occurrences of given category to be kept. Less frequent categories will be deleted.}

\item{absolute}{logical. should the series be transformed to absolute values?}

\item{nIter}{integer. Number of random extractions.}

\item{rotate}{if true windows are extracted from a randomly shifted timeseries. This is better TRUE.}

\item{extract_from}{Either "all", or "remaining". Should new extractions be done randomly from the whole stream or only on non-categorized parts.}

\item{summarizingFunction}{String. the function used to summarize each epoch's time-series to a single value. Suggested: median.}

\item{parameterFunction}{String. the function used to summarize all epochs's single values. Suggested: median.}

\item{xlim}{either "auto" or a numeric vector of length 2 to set the plot scale.}
}
\description{
given a well formatted DyadExperiment object with some time series (e.g. synchronization or physiology) and DyadCategory data
it cuts the series according to the categories intervals, extracts random permutations of the stream, and calculates p-values and effect sizes
for the observed central values compared to random ones.
}
