% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract.R
\name{catExtract}
\alias{catExtract}
\title{Extract summary data from categorial windows
the function first extracts all occurrences of a given category, combines them in a
single data.frame, then aggregates it according to specified rules.}
\usage{
catExtract(experiment, category = "PACS", by, FUN = mean, ...)
}
\arguments{
\item{experiment}{}

\item{category}{}

\item{by}{a character vector of one or more between "session","dyad","group" and the category's column names}

\item{FUN}{a function to compute the summary statistics which can be applied to all data subsets}

\item{...}{arguments passed to FUN}
}
\description{
Extract summary data from categorial windows
the function first extracts all occurrences of a given category, combines them in a
single data.frame, then aggregates it according to specified rules.
}
