% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_extraction.R
\name{catSummary}
\alias{catSummary}
\title{Prints a summary of all epochs}
\usage{
catSummary(x, diaryName)
}
\arguments{
\item{x}{a DyadExperiment or DyadSession object.}

\item{diaryName}{string. The name of the Diary to be summarized}
}
\value{
nothing.
}
\description{
Prints a summary of all epochs
}
