% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_CCFBest.R
\name{ccfBest}
\alias{ccfBest}
\title{Title}
\usage{
ccfBest(x, signals = "all", lagSec, winSec, incSec, accelSec,
  weight_type = c("center", "free", "off"), simplify = T)
}
\arguments{
\item{x}{}

\item{signals}{}

\item{lagSec}{}

\item{winSec}{}

\item{incSec}{}

\item{accelSec}{}

\item{weight_type}{}

\item{simplify}{}
}
\description{
Title
}
