% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_CCFBest.R
\name{ccfBestSlope}
\alias{ccfBestSlope}
\title{ccfBest function but applied on the slope
eventually a moving average is further performed on data to mimic Marci et al 2007 procedure.}
\usage{
ccfBestSlope(x, signals = "all", lagSec, winSec, incSec, accelSec, MAwin,
  MAinc, weight_type = c("center", "free", "off"), simplify = T)
}
\arguments{
\item{x}{}

\item{signals}{}

\item{lagSec}{}

\item{winSec}{}

\item{incSec}{}

\item{accelSec}{}

\item{weight_type}{}

\item{simplify}{}
}
\description{
ccfBest function but applied on the slope
eventually a moving average is further performed on data to mimic Marci et al 2007 procedure.
}
