% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{ccfQuantile}
\alias{ccfQuantile}
\title{Title}
\usage{
ccfQuantile(EXP, signal = "SC", lag = "bestCCF", bySession = T,
  sessionFUN = "mean")
}
\arguments{
\item{EXP}{}

\item{signal}{}

\item{lag}{}

\item{bySession}{}

\item{sessionFUN}{}
}
\description{
Title
}
