% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_classes.R
\name{classAttr}
\alias{classAttr}
\alias{classAttr<-}
\title{Safe Object Attribute Lists}
\usage{
classAttr(x)

classAttr(x) <- value
}
\arguments{
\item{x}{an object}

\item{value}{an appropriate named list of attributes, or NULL.}
}
\value{
asd
}
\description{
a wrapper for \code{\link[base]{attributes}} which accesses and sets an object's attributes
with the exception of protected ones, currently: \code{c("names", "comment", "dim", "dimnames", "row.names", "tsp")}
}
\examples{
a <- list("a"=1,"b"=2,"c"=3)
attr(a, "foo") <- "bar"
#extracts as well names
attributes(a) 
#extracts only foo
classAttr(a) 
#overwrites only foo
classAttr(a) <- list("bar"="foo") 

}
