% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_get_set.R
\name{cloneAttr}
\alias{cloneAttr}
\title{clone attributes}
\usage{
cloneAttr(from, to)
}
\arguments{
\item{from}{}

\item{to}{}
}
\value{
the object 'to' with the attributes of 'from'. This does not overwrites
protected attributes: "names", "dim", "dimnames", "row.names", "tsp"
}
\description{
clone attributes
}
