% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_videoframes.R
\name{drawFrames}
\alias{drawFrames}
\title{This library provides tools to overlay signals on video}
\usage{
drawFrames(
  session,
  signals,
  sync,
  WIN,
  inputDir,
  outputDir,
  idOrder = c(F, F, T),
  idSep = "_",
  w,
  h,
  frameRate,
  layout = c(1),
  col_sync,
  ffmpeg.query,
  ffmpeg.path,
  multi.core = TRUE,
  speed = 1,
  keepAudio = TRUE,
  ...
)
}
\arguments{
\item{session}{}

\item{signals}{}

\item{sync}{}

\item{WIN}{}

\item{outputDir}{}

\item{w}{}

\item{h}{}

\item{frameRate}{}

\item{layout}{}

\item{col_sync}{}

\item{...}{}

\item{ffmpeg}{override the default ffmpeg instruction. Use {IN} to mark the dynamic
original video location. Not yet implemented}
}
\description{
to allow for multiple signals, everything must be a list of length n
so if I have e.g. SC, HR and MEA, I must provide 3 dyadsignals, 3 y positions, etc
}
\details{
Directory organization
Each of these endeavors require:
- json file of the original acquisition
- csv version for DyadSync import
- video of the session in a suitable format
- folders with all the frames
- output video file
THUS it would make sense to have a folder for each session instead of a folder
for each 'type' of data
Draws video frames showing dyadic signals
}
\examples{
quindi il tempo tra 00:00 e start(signal) va riempito di frame bianchi
se start(signal) è == 00:00 bisogna aggiungere WIN/2 sample perché il primo sample
deve comparire al centro della finestra
}
