% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_extraction.R
\name{epochSeries}
\alias{epochSeries}
\title{Extract series chunks corresponding to epochs}
\usage{
epochSeries(
  x,
  signal,
  sync,
  series,
  category,
  categoryIndex,
  mergeEpochs = FALSE,
  artefact.rm = TRUE,
  shift_start = 0,
  shift_end = 0
)
}
\arguments{
\item{x}{a DyadExperiment object}

\item{signal}{string. The name of a DyadSignal present in x}

\item{sync}{If missing, series is searched in the signal (s1, s2, ...). If specified, series is searched within a sync object (PMBdev, CCFBest, ...).}

\item{series}{}

\item{category}{}

\item{categoryIndex}{}

\item{mergeEpochs}{}

\item{artefact.rm}{}

\item{shift_start}{integer in seconds (e.g. -5 shifts all epoch starts by 5 seconds before )}

\item{shift_end}{integer in seconds (e.g. 7 shifts all epoch end by 7 seconds later )}
}
\description{
epochSeries cycles through all sessions in an experiment.
For each session extracts a series, then cuts it
according to categories epochs
}
