% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{expApply}
\alias{expApply}
\title{Title
applies a function on each signal of an experiment.}
\usage{
expApply(experiment, signals = "all", FUN, ...)
}
\arguments{
\item{experiment}{}

\item{signals}{}

\item{FUN}{}

\item{...}{}
}
\description{
Title
applies a function on each signal of an experiment.
}
