% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_PMBest.R
\name{expPPsync}
\alias{expPPsync}
\title{Title}
\usage{
expPPsync(experiment, signals = "all", lagSec = 7, sgol_p = 2,
  sgol_n = 25, weightMalus = 30)
}
\arguments{
\item{experiment}{}

\item{signals}{}

\item{lagSec}{}
}
\description{
Title
}
