% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolescense.R, R/rIP_extract_dev.R
\name{catExtractLong}
\alias{catExtractLong}
\alias{extractEpochs}
\title{Here goes deprecated and obsolete functions calls, that will eventually get deleted}
\usage{
catExtractLong(x, signal = "", epochStream = "", by = "", FUN = "")

extractEpochs(
  experiment,
  signal,
  sync,
  stream,
  category,
  categoryIndex,
  by,
  epochStreamName,
  ...
)
}
\arguments{
\item{signal}{}

\item{by}{currently unused. In future will be used to split the data by experimental group, participant, or any other relevant condition}

\item{FUN}{}

\item{experiment}{}

\item{sync}{}

\item{stream}{}

\item{category}{the category used to}

\item{epochStreamName}{deprecated. use sync, streamkey, category}

\item{...}{}
}
\description{
This function extracts the selected epochs from every session of a "DyadExperiment" object and puts them in
 a simple list. Categories must be created with epochStream() beforehand.
}
