% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{flexMA}
\alias{flexMA}
\title{Title}
\usage{
flexMA(x, win, remove = F)
}
\arguments{
\item{x}{}

\item{win}{if x has a frequency, it will be the window size in seconds. Otherwise in samples.}

\item{remove}{}
}
\description{
Title
}
\examples{
a = c(rep(0,20),rep(5,20),rep(10,20),rep(20,20))
plot(a)
lines(flexMA(a,10,F))
lines(flexMA(a,20,F),col=3)
b=DyadStream(a,"test",col=2,frequency = 10, start=0)
lines(flexMA(b,2,F),col=color(b),lty=3)
sig=list("s1" = b, s2 = b)
class(sig) ="DyadSignal" 
lines(flexMA(sig$s1,2,F),col="skyblue",lwd=2)
}
