% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_CCFBest.R
\name{getCCF}
\alias{getCCF}
\title{Title}
\usage{
getCCF(signal, lag, col = "red", lty = 2, lwd = 2, interpolate = F)
}
\arguments{
\item{signal}{}

\item{lag}{}

\item{col}{}

\item{lty}{}

\item{lwd}{}

\item{interpolate}{}
}
\description{
Title
}
