% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_classes.R
\name{hasSync}
\alias{hasSync}
\title{hasSync
Recursive function which checks if any of the nested object(s) (typically a DyadSignal)
contains at least one children of a class defined in the global constant SYNC_CLASSES}
\usage{
hasSync(x)
}
\arguments{
\item{x}{}
}
\description{
hasSync
Recursive function which checks if any of the nested object(s) (typically a DyadSignal)
contains at least one children of a class defined in the global constant SYNC_CLASSES
}
