% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{is.color}
\alias{is.color}
\title{Checks if valid color representation}
\usage{
is.color(x)
}
\arguments{
\item{x}{a vector of characters}
}
\value{
a vector of named logicals
}
\description{
Checks if valid color representation
}
