% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{lead0}
\alias{lead0}
\title{Formats number with leading zeros}
\usage{
lead0(x, width = 2, digits = 0)
}
\arguments{
\item{x}{}

\item{width}{number of leading zeros}

\item{digits}{number of decimal digits}
}
\description{
Formats number with leading zeros
}
