% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{lowpass}
\alias{lowpass}
\title{lowPass}
\usage{
lowpass(x, cut)
}
\arguments{
\item{x}{a DyadStream or a ts object}

\item{cut}{the cutoff frequency in herz}
}
\value{
a DyadStream or a ts object
}
\description{
lowPass
}
