% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_CCFBest.R
\name{marciIndex}
\alias{marciIndex}
\title{Marci's concordance index}
\usage{
marciIndex(a)
}
\arguments{
\item{a}{a vector of cross-correlations, or a list of them}
}
\description{
Marci's concordance index
}
\details{
see Marci et al. (2007)
}
