% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_IO_export.R
\name{megaExpExport}
\alias{megaExpExport}
\title{Title}
\usage{
megaExpExport(
  dirPath,
  experiment,
  signals = "all",
  CCF = T,
  onlyBest = F,
  original = T
)
}
\arguments{
\item{dirPath}{}

\item{experiment}{}

\item{signals}{}

\item{CCF}{}

\item{onlyBest}{}

\item{original}{}
}
\value{

}
\description{
Title
}
