% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_random.R
\name{nCombo}
\alias{nCombo}
\title{Calculate the number of possible pseudo-dyads}
\usage{
nCombo(n_dyads, max_combo = 10000)
}
\arguments{
\item{n_dyads}{}

\item{max_combo}{}
}
\description{
Calculate the number of possible pseudo-dyads
}
