% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_sync_classes.R
\name{newAMICo}
\alias{newAMICo}
\title{an AMICo analysis contains 3 objects:
a sync, and lag vector, containing synchrony and lag at the same frequency
of the original time series
and xBest, which is a table with the matched peaks positions and their similarity
also there are many attributes that save how the analysis was run}
\usage{
newAMICo(sync, lag, xBest, args)
}
\description{
an AMICo analysis contains 3 objects:
a sync, and lag vector, containing synchrony and lag at the same frequency
of the original time series
and xBest, which is a table with the matched peaks positions and their similarity
also there are many attributes that save how the analysis was run
}
