% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{nwin}
\alias{nwin}
\title{Calculate number of windows in a time-series with a given duration}
\usage{
nwin(
  x,
  winSec,
  incSec,
  SR = frequency(x),
  return = c("number", "all"),
  flex = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{either the duration in seconds of a time-series, or a rats or ts object}

\item{winSec}{the width of each window, in seconds}

\item{incSec}{the amount of increment between each window (incSec == winSec gives non-overlapping windows)}

\item{SR}{the number of samples per second (i.e. frequency)}

\item{return}{either "number", which returns the number of windows, or "all" which returns a data.frame with the start and end of each window.}

\item{flex}{if true the first and last windows are stretched so to have exactly length(x)/inc resulting windows}

\item{verbose}{should all the windows be printed?}
}
\description{
Calculate number of windows in a time-series with a given duration
}
