% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_extraction.R
\name{oneDiary}
\alias{oneDiary}
\title{Join together all sessions of a given Diary}
\usage{
oneDiary(exp, diaryName)
}
\arguments{
\item{exp}{a DyadExperiment object}

\item{diaryName}{string. The name of the Diary to be exported}
}
\value{
a dataframe containing all sessions of the selected Diary
}
\description{
Join together all sessions of a given Diary
}
