% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals_processing.R
\name{peakFinder}
\alias{peakFinder}
\title{Finds peaks and throughs features in a time series}
\usage{
peakFinder(
  x,
  mode = c("peaks", "valleys", "both"),
  sgolay_args = list(),
  FIR_args = list(),
  correctionRangeSeconds = 0,
  minPeakAmplitude = 0,
  sgol_p,
  sgol_n
)
}
\arguments{
\item{x}{a rats, ts, or a numeric vector.}

\item{mode}{should the function return only 'peaks', 'valleys', or 'both'?}

\item{sgolay_args}{a named list of arguments to be passed to \link[signal]{sgolay}) Savitzky-Golay smoothing filter.
'p' and 'n' must be present, representing the filter order and length (this must be an odd value). If empty, the filtering is disabled.}

\item{FIR_args}{a named list of arguments to be passed to \link[DyadSignal]{FIR}) filtering function.
'cut' and 'type' must be present, representing . If empty, the filtering is disabled.}

\item{correctionRangeSeconds}{the half range in which the real maximum/minimum
value should be searched, in seconds.
around the derivative shift. Should be less then the periodicity of the
signal.  0.5 is good for skin conductance.}

\item{minPeakAmplitude}{the minimum delta from valley to peak for detection.
Skin conductance traditionally uses 0.05uS, or 0.03uS}

\item{sgol_p}{OBSOLETE parameter: smoothing filter order of the sgolay filter.}

\item{sgol_n}{OBSOLETE parameter: smoothing filter length (must be odd).}
}
\value{
a list of:
\describe{
  \item{bool}{a logical vector of the same length of x with TRUE
   value corresponding to a match
   }
  \item{samples}{the position of features relatively to x's index}
  \item{x}{temporal coordinates of the features (if x has a frequency attribute)}
  \item{type}{a character vector defining for each "samples" value if its a 'p' peak or a 'v' valley (trough)}
  \item{y}{The value of x correspoding to the detected features}
  \item{amp}{the positive (v->p) and negative (p->v) amplitudes of the detected features}
  \item{i}{The absolute position of the features along the signal. The values
     are independent from the 'mode' argument, allowing interaction between
     different calls of the function.}
  
  
}
}
\description{
Finds peaks and throughs features in a time series
}
\details{
The function has a relatively simple yet robust implementation:
after a Sgolay smoothing, the first derivative's sign changes are detected.
Then, the actual minima or maxima is looked for in a given range.
NOTE: the function ensures that no consecutive throughs or peaks are returned.
}
