% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_graphics.R
\name{plot.DyadSignal}
\alias{plot.DyadSignal}
\title{Plot a DyadSignal object}
\usage{
\method{plot}{DyadSignal}(x, sync = NA, rescale = c("none", "win", "fixed"), ...)
}
\arguments{
\item{x}{a DyadSignal object}

\item{sync}{Either NA or a string pointing to the name of a PMBest or CCFBest object}

\item{...}{}
}
\description{
Plot a DyadSignal object
}
