% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotAMICo}
\alias{plotAMICo}
\title{Plots a window of signal and the AMICo algorithm results}
\usage{
plotAMICo(signal, start, end, syncName = "AMICo")
}
\arguments{
\item{signal}{object of class DyadSignal}

\item{start}{start of plotting window. "mm:ss" or a number of seconds}

\item{end}{end of plotting window. "mm:ss" or a number of seconds}
}
\description{
Plots a window of signal and the AMICo algorithm results
}
