% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{print.rats}
\alias{print.rats}
\title{Print a rats object}
\usage{
\method{print}{rats}(x, vals = 20, digits = 4)
}
\arguments{
\item{x}{}

\item{vals}{numeric. How many values should be printed?}

\item{digits}{numeric. How many digits should be used to approximate values}
}
\description{
Print a rats object
}
