% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{rangeRescale}
\alias{rangeRescale}
\title{rangeRescale
this extremely useful function rescales a numeric vector to a new range}
\usage{
rangeRescale(x, rangeMin, rangeMax, xmin = min(x, na.rm = T), xmax = max(x,
  na.rm = T))
}
\arguments{
\item{x}{}

\item{rangeMin, rangeMax}{il minimo e il massimo della nuova scala}

\item{xmin, xmax}{il minimo e il massimo della scala originale.
If xmin or xmax are missing the min or max of x are used instead}
}
\description{
rangeRescale
this extremely useful function rescales a numeric vector to a new range
}
