% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_IO_categories.R
\name{readCategories}
\alias{readCategories}
\title{Title}
\usage{
readCategories(
  path,
  startCol,
  endCol,
  catName,
  namefilt = NA,
  removeSec = 0,
  idOrder = c("id", "session", "group"),
  idSep = "_",
  ...
)
}
\arguments{
\item{startCol}{}

\item{endCol}{}

\item{catName}{}

\item{namefilt}{}

\item{removeSec}{}

\item{idOrder}{}

\item{idSep}{}

\item{...}{}

\item{dirPath}{}
}
\description{
Title
}
