% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_IO.R
\name{readDyadExperiment}
\alias{readDyadExperiment}
\title{Title}
\usage{
readDyadExperiment(path, start = 0, end = FALSE, paCol, teCol, signalNames,
  sampRate, pairBind = F, winTerpolate = list(winSec = NULL, incSec = NULL),
  namefilt = NA, sumCols = FALSE, idOrder = c(), idSep = "_", ...)
}
\arguments{
\item{path}{}

\item{start}{a vector of integers which specify the time (in seconds) of the first observation of each file.
Note that the default value (0) is different from the default value of \code{\link[stats]{ts}}.}

\item{end}{a vector of integers which specify the time (in seconds) of the last observation of each file.
If the data is longer or shorter, remove tail
 or pads with zeroes. Useful to cut signals when a session is over or to equalize lengths. The default value,
 FALSE, keeps the original length of the signal}

\item{paCol}{}

\item{teCol}{}

\item{signalNames}{}

\item{sampRate}{}

\item{pairBind}{}

\item{winTerpolate}{}

\item{namefilt}{}

\item{sumCols}{}

\item{idOrder}{}

\item{idSep}{}

\item{...}{}
}
\description{
Title
}
