% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{resample}
\alias{resample}
\title{resample
A simple wrapper for approx, used to decimate or upsample (by linear interpolation) a time series}
\usage{
resample(x, newSampRate, ...)
}
\arguments{
\item{x}{A time-series object}

\item{newSampRate}{the new frequency}

\item{...}{further options to be passed to approx}
}
\value{
a resampled time-series with the same start of the original time series.
}
\description{
resample
A simple wrapper for approx, used to decimate or upsample (by linear interpolation) a time series
}
