% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{rescaleByWin}
\alias{rescaleByWin}
\title{Approx by window
when you have a value for each window (even overlapping ones) and you
want to transform the values to a time series with the original sampling rate}
\usage{
rescaleByWin(x, WIN, newMin, newMax, cores = 1)
}
\arguments{
\item{x}{numeric representation of a time-series. Ideally a rats object.}

\item{WIN}{numeric. Size of the window in the original series time unit (e.g. seconds)}

\item{newMin}{}

\item{newMax}{}

\item{cores}{}

\item{a}{a numerical vector, resulting from a windowing}

\item{winSec}{the size of the window used to calculate a, in seconds}

\item{incSec}{the increment of each window used to calculate a, in seconds}

\item{SR}{the number of samples per second of the destination time-series}

\item{midPoints}{if true, the value are considered to be at the center of the window, and the result is padded right by half window}
}
\description{
This function allows to visualize the short-term dynamics of a
time series using the full vertical resolution of a screen
}
\examples{
Rescale every sample of a time series over a moving window 

}
