% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_get.R
\name{sampRate}
\alias{sampRate}
\title{#' @export
uniqueId.DyadExperiment = function(x){
  `Group Ids` = sapply(x, attr, "groupId")
  table(`Group Ids`)
}}
\usage{
sampRate(x)
}
\description{
#' @export
uniqueId.DyadSession = function(x){
  attr(x,"groupId")
}
}
