% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{selectSignals}
\alias{selectSignals}
\title{Title
This function allows to subset only specific signals from an experiment}
\usage{
selectSignals(experiment, signals = c("test1", "test2"))
}
\arguments{
\item{experiment}{}

\item{signals}{}
}
\description{
Title
This function allows to subset only specific signals from an experiment
}
