% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{setArtefacts}
\alias{setArtefacts}
\title{setArtefacts
setArtefacts legge una tabella o lista con le componenti chiamate esattamente: 'session', 'start' ed 'end'
e mette ad NA tali sequenze nel segnale specificato.
end può anche essere una stringa che contiene le parole 'fine' o 'end'}
\usage{
setArtefacts(x, startEnd, signal = "SC")
}
\arguments{
\item{x}{}

\item{startEnd}{}

\item{signal}{}
}
\description{
setArtefacts
setArtefacts legge una tabella o lista con le componenti chiamate esattamente: 'session', 'start' ed 'end'
e mette ad NA tali sequenze nel segnale specificato.
end può anche essere una stringa che contiene le parole 'fine' o 'end'
}
