% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete_bwcompatibility.R
\name{signalDecimate}
\alias{signalDecimate}
\title{#vecchio "get"}
\usage{
signalDecimate(signal, newSampRate)
}
\description{
#vecchio "get"
}
\details{
#' @export
frequency.DyadSignal = function(x){attr(x,"SR")}
}
