% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{signalMerge}
\alias{signalMerge}
\title{Title
this function adds the signals from two experiments}
\usage{
signalMerge(exp1, exp2)
}
\arguments{
\item{exp1}{}

\item{exp2}{}
}
\description{
Title
this function adds the signals from two experiments
}
