% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract.R
\name{streamExtract}
\alias{streamExtract}
\title{Extract summary data from a stream
The function iterates over each session of an experiment and applies a summarizing function on a given stream (e.g. mean).
Multiple output are possible (eg. using the quantile function)}
\usage{
streamExtract(experiment, signal, sync = "none", streamKey = c("s1", "s2",
  "time", "sync", "lag", "zero"), FUN = mean, ...)
}
\arguments{
\item{experiment}{}

\item{signal}{}

\item{sync}{}

\item{streamKey}{}

\item{FUN}{}

\item{...}{}
}
\description{
Extract summary data from a stream
The function iterates over each session of an experiment and applies a summarizing function on a given stream (e.g. mean).
Multiple output are possible (eg. using the quantile function)
}
