% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{timeMaster}
\alias{timeMaster}
\title{timeMaster
timeMaster allows to:
transform time from different formats to different formats.
add amounts of time to baseTime expressed in different formats.
This is verified. And useful. For reasons. :-D}
\usage{
timeMaster(
  baseTime,
  out = c("auto", "hour", "min", "sec"),
  add = 0,
  baseSep = "[:,;,\\\\,',\\",\\\\-]",
  digits = c("auto")
)
}
\arguments{
\item{baseTime}{numeric representig seconds, or a string in the h:m:s or m:s format}

\item{out}{the output format "auto", "hour" for hh:mm:ss, "min" for "mm:ss"}

\item{add}{a time to be added (or subtracted if negative) from basetime. In any format accepted by timeMaster}

\item{baseSep}{the separator dividing hh mm ss}

\item{digits}{the number of digits of the fractional part of seconds}
}
\description{
timeMaster
timeMaster allows to:
transform time from different formats to different formats.
add amounts of time to baseTime expressed in different formats.
This is verified. And useful. For reasons. :-D
}
