% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_graphics.R
\name{tsAxis}
\alias{tsAxis}
\title{Easy axes for ts objects}
\usage{
tsAxis(x, side, every = 1, out = c("hour", "min", "sec"), las = 1, ...)
}
\arguments{
\item{x}{a ts object}

\item{side}{side of the axie}

\item{every}{interval between ticks, in seconds}

\item{out}{format of the time, one of "hour","min","sec" for respectively the formats
hh:mm:ss, mm:ss, ss.}

\item{las}{'label axis style' sets the labels orientation. See ?par for details}

\item{...}{further arguments passed to axis}
}
\value{

}
\description{
Easy axes for ts objects
}
