% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing_ts.R
\name{tse}
\alias{tse}
\title{Time Series End
this returns the end of a ts object in time units}
\usage{
tse(x)
}
\arguments{
\item{x}{a ts object}
}
\description{
Time Series End
this returns the end of a ts object in time units
}
