% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing_ts.R
\name{tss}
\alias{tss}
\title{Time Series Start
this returns the start of a ts object in time units}
\usage{
tss(x)
}
\arguments{
\item{x}{a ts object}
}
\description{
Time Series Start
this returns the start of a ts object in time units
}
