% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_main.R
\name{window.DyadSignal}
\alias{window.DyadSignal}
\title{Time Windows}
\usage{
\method{window}{DyadSignal}(x, duration = NULL, ...)
}
\arguments{
\item{x}{}

\item{duration}{an alternative specification to end}

\item{...}{}
}
\description{
Time Windows
}
