% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{window.rats}
\alias{window.rats}
\title{Subsets rats by time
Intervals always include the 'start' and exclude the 'end' values.
Windows can alternatively be specified through start, end, or one among
start and end and a duration.}
\usage{
\method{window}{rats}(x, start, end, duration)
}
\arguments{
\item{x}{}

\item{start}{numeric. Start of the window}

\item{end}{numeric. End of the window}

\item{duration}{numeric. The duration of the window.}
}
\value{
a subset rats object
}
\description{
Subsets rats by time
Intervals always include the 'start' and exclude the 'end' values.
Windows can alternatively be specified through start, end, or one among
start and end and a duration.
}
\examples{
x = rats(1:100, start=0, frequency = 10)


window(x, start=2, duration=2)

#Replace values within a window
window(x, start=0.21, end=0.5) = 99 


-5

}
