% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_AMICo2_RC.R
\name{AMICo}
\alias{AMICo}
\title{Peak matching similarity of signals}
\usage{
AMICo(
  experiment,
  signal,
  lagSec = 6,
  weightMalus = 50,
  match_threshold = 0.1,
  minSizeSec = 4,
  maxPeakDuration = "hrt",
  interval_sec = 10,
  algorithm = c("v2RC0.4"),
  outputName = "AMICo",
  sgol_p = 2,
  sgol_n = 25,
  correctionRangeSeconds,
  minPeakAmplitude
)
}
\arguments{
\item{experiment}{}

\item{signal}{the name of a DyadSignal contained in experiment}

\item{lagSec}{}

\item{weightMalus}{weightmalus è la percentuale di malus per il lag più estremo. Es, con weightMalus= 20 e r = 1 al massimo lag, la trasformazione diventa r' = 0.8}

\item{match_threshold}{a value between 0 and 1, specifying the similarity threshold to assign a peak-peak match.}

\item{minSizeSec}{the smallest allowed window size. very small windows may artificially inflate synchrony.}

\item{maxPeakDuration}{}

\item{interval_sec}{in AMICO2, how long should unmatched sequences be?}

\item{algorithm}{character. The version of AMICo algorithm.}

\item{...}{Further arguments passed to peakFinder. Specifically, 
sgol_p, sgol_n, correctionRangeSeconds and minPeakAmplitude must be set
according to the signal type.}
}
\description{
Peak matching similarity of signals
}
\details{
Good values for skin conductance are correctionRangeSeconds = 0.5,
minPeakAmplitude = 0.05.

The difference in algorithms can be found in the changelog. Basically between
1.0 and 1.1 there is a change in peakFinder to avoid detecting micro-peaks in generally flat areas.

The difference with 2.0 is much more substantial. Instead of calculating
similarities twice, in v2 the same approach is used to choose the best peak-to-peak
associations and to report the final values. Also exclusively valley-peak-valley
or valley-peak-halfrecovery sequences are used for the calculation.
Unmatched sequences are evenly split at interval_sec intervals.

XBEST guide:
row: the peak number of s1 wich was matched
col: the peak number of s2
similarity: some similarity computation
lag: the delta between the sample of p1 and p2
a1, p1, b1: sample position of onset, peak, and end of a feature
a2, p2, b2: the same for s2
a, b: the average start and end between s1 and s2
ta, tb: the time corresponding to a and b
}
