% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_PMBest.R
\name{AMICo}
\alias{AMICo}
\title{pmBest}
\usage{
AMICo(
  experiment,
  signal = "all",
  lagSec = 20,
  sgol_p = 2,
  sgol_n = 25,
  weightMalus = 20,
  match_threshold = 0.25,
  minSizeSec = 1,
  algorithm = c("v1.1", "v1.0"),
  outputName = "AMICo",
  correctionRangeSeconds = 0.5,
  minPeakDelta
)
}
\arguments{
\item{experiment}{}

\item{signal}{the name of a DyadSignal contained in experiment}

\item{lagSec}{}

\item{sgol_p}{}

\item{sgol_n}{}

\item{weightMalus}{weightmalus è la percentuale di malus per il lag più estremo. Es, con weightMalus= 20 e r = 1 al massimo lag, la trasformazione diventa r' = 0.8}

\item{match_threshold}{a value between 0 and 1, specifying the similarity threshold to assign a peak-peak match.}

\item{minSizeSec}{the smallest allowed window size. very small windows may artificially inflate synchrony.}

\item{correctionRangeSeconds}{parameter passed to peakFinder, seconds of manual correction for derivative based peak-finding}

\item{minPeakDelta}{parameter passed to peakFinder, minimum valley-peak delta to detect a peak}

\item{scaledCorrelation}{logical. If TRUE, correlations will be scaled using the whole signals range, and be much more robust}
}
\description{
pmBest
}
