% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract.R
\name{catApply}
\alias{catApply}
\title{Title}
\usage{
catApply(x, FUN, signal = "SC", sync = c("none", SYNC_CLASSES),
  streamKey = c("s1", "s2", "time", "sync", "lag", "zero"),
  category = "PACS", ...)
}
\arguments{
\item{x}{}

\item{FUN}{}

\item{signal}{}

\item{sync}{}

\item{streamKey}{}

\item{category}{}

\item{...}{arguments passed to FUN}

\item{column}{}
}
\description{
Title
}
