% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_classes.R
\name{CCFBest}
\alias{CCFBest}
\title{SYNC_CLASSES = c("CCFBest","PMBest")}
\usage{
CCFBest(sync, lag, ccf_matrix, lagSec, winSec, incSec, accelSec,
  weight_type, MAwin, MAinc, slopes)
}
\description{
SYNC_CLASSES = c("CCFBest","PMBest")
}
