% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_CCFBest.R
\name{ccfBest}
\alias{ccfBest}
\title{Title}
\usage{
ccfBest(x, signals = "all", lagSec, winSec, incSec, accelSec,
  weight_type = c("center", "free"), simplify = T)
}
\arguments{
\item{signals}{}

\item{lagSec}{}

\item{winSec}{}

\item{incSec}{}

\item{accelSec}{}

\item{simplify}{}

\item{experiment}{}

\item{weight}{}

\item{interpolate}{}
}
\description{
Title
}
\examples{

}
