% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_classes.R
\name{DyadStream}
\alias{DyadStream}
\title{DyadStream}
\usage{
DyadStream(stream, name, ...)
}
\arguments{
\item{stream}{a ts object or numeric vector}

\item{name}{}

\item{...}{arguments passed to ts(), typicaly frequency and start. Only used if stream is NOT a ts object.}
}
\description{
DyadStream
}
