% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_extraction.R
\name{epochSeries}
\alias{epochSeries}
\title{Extract series chunks corresponding to epochs}
\usage{
epochSeries(
  x,
  signal,
  sync,
  series,
  diary,
  category,
  mergeEpochs = FALSE,
  artefact.rm = TRUE,
  shift_start = 0,
  shift_end = 0,
  target = c("epoch", "before", "after"),
  summaryFUN = "median",
  ...
)
}
\arguments{
\item{x}{a DyadExperiment object}

\item{signal}{string. The name of a DyadSignal present in x}

\item{sync}{If missing, series is searched in the signal (s1, s2, ...). If specified, series is searched within a sync object (PMBdev, CCFBest, ...).}

\item{series}{}

\item{diary}{}

\item{category}{}

\item{mergeEpochs}{}

\item{artefact.rm}{}

\item{shift_start}{numeric. Seconds to shift all start values (e.g. -5 shifts all epoch starts by 5 seconds before )}

\item{shift_end}{numeric. Seconds to shift all end values (e.g. 7.5 shifts all epoch end by 7.5 seconds later )}

\item{target}{one of "epoch", "before", "after". Determines whether to extract the actual epochs or only what precedes or follows it.
If it's not set to "epoch", shift_start or shift_end must be specified.}

\item{summaryFUN}{}

\item{...}{}
}
\description{
epochSeries cycles through all sessions in an experiment.
For each session extracts a series, then cuts it
according to categories epochs
}
