% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract_dev.R
\name{epochStream}
\alias{epochStream}
\title{Extract stream chunks corresponding to epochs}
\usage{
epochStream(
  x,
  signal = "SC",
  sync = "PMBest",
  streamKey = "sync",
  category,
  groupIndex,
  mergeEpochs = FALSE,
  artefact.rm = TRUE,
  shift_start = 0,
  shift_end = 0
)
}
\arguments{
\item{x}{}

\item{signal}{asd}

\item{sync}{}

\item{streamKey}{}

\item{category}{}

\item{groupIndex}{}

\item{mergeEpochs}{}

\item{artefact.rm}{}

\item{shift_start}{integer in seconds (e.g. -5 shifts all epoch starts by 5 seconds before )}

\item{shift_end}{integer in seconds (e.g. 7 shifts all epoch end by 7 seconds later )}
}
\description{
epochStream cycles through all sessions in an experiment.
For each session extracts a stream, then cuts it
according to categories epochs
}
