% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract_dev.R
\name{epochStream}
\alias{epochStream}
\title{Extract stream chunks corresponding to epochs}
\usage{
epochStream(x, signal = "SC", sync = "PMBest", streamKey = "sync",
  category, groupIndex, mergeEpochs = F, artefact.rm = "T")
}
\arguments{
\item{x}{}

\item{signal}{}

\item{sync}{}

\item{streamKey}{}

\item{category}{}

\item{groupIndex}{}

\item{artefact.rm}{}
}
\description{
epochStream cycles through all sessions in an experiment.
For each session extracts a stream, then cuts it
according to categories epochs
}
