% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract.R
\name{epochStreamApply}
\alias{epochStreamApply}
\title{Title
epochStreamApply cycles through all sessions in an experiment. For each session extracts a stream, then cuts it
according to categories epochs, and applies a function to each epoch.}
\usage{
epochStreamApply(
  x,
  FUN,
  signal = "SC",
  sync = "none",
  streamKey = c("s1", "s2", "time", "sync", "lag", "zero"),
  category = "PACS",
  ...
)
}
\arguments{
\item{x}{}

\item{FUN}{}

\item{signal}{}

\item{sync}{either "none" or the name of a DyadSignal component, such as CCFBest or PMBest}

\item{streamKey}{}

\item{category}{}

\item{...}{arguments passed to FUN}

\item{column}{}
}
\value{
the updated DyadExperiment, or single DyadSession object.
}
\description{
Title
epochStreamApply cycles through all sessions in an experiment. For each session extracts a stream, then cuts it
according to categories epochs, and applies a function to each epoch.
}
