% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{expApply}
\alias{expApply}
\title{THIS IS OLD
USE signalFilter instead
applies a function on each signal of an experiment.}
\usage{
expApply(experiment, signals = "all", FUN, ...)
}
\arguments{
\item{experiment}{}

\item{signals}{either the string "all" or a vector of signal names.
If 'all' the function will be applied on all DyadSignal objects. To apply on epochs,
and other objects, the names must be specified}

\item{FUN}{a function passed to a lapply call on a list of DyadSignals}

\item{...}{further arguments passed to lapply}
}
\description{
THIS IS OLD
USE signalFilter instead
applies a function on each signal of an experiment.
}
\details{
the FUN can be in the generic form function(x, ...){ } where x is each DyadSignal object
among the selected signals, in each session.
}
