% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_PMBest.R
\name{expPPsync}
\alias{expPPsync}
\title{expPPsync}
\usage{
expPPsync(experiment, signals = "all", lagSec = 7, sgol_p = 2,
  sgol_n = 25, weightMalus = 30)
}
\arguments{
\item{experiment}{}

\item{signals}{}

\item{lagSec}{}

\item{sgol_p}{}

\item{sgol_n}{}

\item{weightMalus}{weightmalus è la percentuale di malus per il lag più estremo. Es, con weightMalus= 20 e r = 1 al massimo lag, la trasformazione diventa r' = 0.8}
}
\description{
expPPsync
}
