% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract_dev.R
\name{extractEpochs}
\alias{extractEpochs}
\alias{catExtractLong}
\title{extract epochs in a simple list
 This function extracts the selected epochs from every session of a "DyadExperiment" object and puts them in
 a simple list.
 In future the 'by' argument will be used to split the data by experimental group, participant, or any other relevant condition.}
\usage{
extractEpochs(
  experiment,
  signal = "SC",
  epochStreamName = "IM_PmdevSync",
  by,
  FUN = mean,
  ...
)

catExtractLong(
  experiment,
  signal = "SC",
  epochStreamName = "IM_PmdevSync",
  by,
  FUN = mean,
  ...
)
}
\description{
extract epochs in a simple list
 This function extracts the selected epochs from every session of a "DyadExperiment" object and puts them in
 a simple list.
 In future the 'by' argument will be used to split the data by experimental group, participant, or any other relevant condition.
}
