% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_sync.R
\name{is.sync}
\alias{is.sync}
\title{Checks if the object is an admitted 'sync analysis type'}
\usage{
is.sync(x)
}
\arguments{
\item{x}{}
}
\description{
Checks if the object is an admitted 'sync analysis type'
}
