% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{movAv}
\alias{movAv}
\title{Title}
\usage{
movAv(x, win, inc, sampRate = frequency(x), remove = F)
}
\arguments{
\item{x}{}

\item{win}{}

\item{inc}{}

\item{sampRate}{}

\item{remove}{}
}
\description{
Title
}
