% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{nwin}
\alias{nwin}
\title{Calculate number of windows in a time-series}
\usage{
nwin(
  lenSec,
  winSec,
  incSec,
  sampRate = 1,
  verbose = F,
  return = c("number", "all")
)
}
\arguments{
\item{lenSec}{a duration to be split in windows, in seconds}

\item{winSec}{the width of each window, in seconds}

\item{incSec}{the amount of increment between each window (incSec == winSec gives non-overlapping windows)}

\item{sampRate}{the number of samples per second (i.e. frequency)}

\item{verbose}{should all the windows be printed?}

\item{return}{either "number", which returns the number of windows, or "all" which returns a data.frame with the start and end of each window.}
}
\value{

}
\description{
Calculate number of windows in a time-series
}
