% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_PMBest.R
\name{peakFinder}
\alias{peakFinder}
\title{Title}
\usage{
peakFinder(x, sgol_p = 2, sgol_n = 25, mode = c("peaks", "valleys",
  "both"), correctionRangeSeconds = 0.5, valid)
}
\arguments{
\item{x}{a ts object or a numeric vector.}

\item{sgol_p}{smoothing filter order.}

\item{sgol_n}{smoothing filter length (must be odd).}

\item{mode}{should the function return only 'peaks', 'valleys', or 'both'?}

\item{correctionRangeSeconds}{the range in which the real maximum/minimum value should be searched, in seconds.
around the derivative shift. Should be less then the periodicity of the signal.  0.5 is good for skin conductance.}

\item{valid}{a logical vector of the same length of x. No peaks/valleys are found where valid is FALSE.}
}
\value{
a list of: "bool" a logical vector of the same length of x with TRUE value corresponding to a match;
"samples" the index of matches in x;
"seconds" the position in seconds of matches (if x is a ts object);
and "type" a charachter vector defining for each "samples" value if its a 'p' peak or a 'v' valley.
}
\description{
Title
}
