% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_Sync_PMBest.R
\name{pmBest}
\alias{pmBest}
\title{pmBest}
\usage{
pmBest(experiment, signals = "all", lagSec = 7, sgol_p = 2, sgol_n = 25,
  weightMalus = 30, match_threshold = 0, minSizeSec = 5,
  algorithm = c("classic", "dev", "sccf"), outputName = "PMBest")
}
\arguments{
\item{experiment}{}

\item{signals}{}

\item{lagSec}{}

\item{sgol_p}{}

\item{sgol_n}{}

\item{weightMalus}{weightmalus è la percentuale di malus per il lag più estremo. Es, con weightMalus= 20 e r = 1 al massimo lag, la trasformazione diventa r' = 0.8}

\item{match_threshold}{a value between 0 and 1, specifying the similarity threshold to assign a peak-peak match.}

\item{minSizeSec}{the smallest allowed window size. very small windows may artificially inflate synchrony.}
}
\description{
pmBest
}
