% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{rangeRescale}
\alias{rangeRescale}
\title{rangeRescale
this function rescales a numeric vector to a new range through linear interpolation}
\usage{
rangeRescale(
  x,
  newMin,
  newMax,
  oldMin = min(x, na.rm = T),
  oldMax = max(x, na.rm = T),
  pres.signs = FALSE
)
}
\arguments{
\item{x}{}

\item{newMin, newMax}{Any two points (typically min and max) of the new scale}

\item{oldMin, oldMax}{The corresponding two points of the original scale.
If oldMin or oldMax are missing the min or max of x are used instead}

\item{pres.sign}{logical. if TRUE the signs of the original data are preserved. newMin and newMax must be opposites (e.g. -1,1)}
}
\description{
rangeRescale
this function rescales a numeric vector to a new range through linear interpolation
}
