% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{rangeRescale}
\alias{rangeRescale}
\title{rangeRescale
this extremely useful function rescales a numeric vector to a new range through linear interpolation}
\usage{
rangeRescale(
  x,
  newA,
  newB,
  oldA = min(x, na.rm = T),
  oldB = max(x, na.rm = T),
  pres.signs = FALSE
)
}
\arguments{
\item{x}{}

\item{newA, newB}{Any two points (typically min and max) of the new scale}

\item{oldA, oldB}{The corresponding two points of the original scale.
If oldA or oldB are missing the min or max of x are used instead}

\item{pres.sign}{logical. if TRUE the signs of the original data are preserved. newA and newB must be opposites (e.g. -1,1)}
}
\description{
rangeRescale
this extremely useful function rescales a numeric vector to a new range through linear interpolation
}
