% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rats.R
\name{rats}
\alias{rats}
\title{State of the file:
this is a v0.2 let's say, in which instead of having a $x $y sructure, I am using
a y and attr(y, 'time') format. this is good because x * 10 is preserved. 
An ideal time series object to represent pyhisio signals should have the
following information:
start_date, 
start, end, duration (milliseconds)
sampling per second (Hz)
windowing history: size, inc, flex}
\usage{
rats(
  data,
  start = 0,
  end,
  duration,
  frequency,
  period,
  windowed = list(winSec = NULL, incSec = NULL, flex = NULL),
  timeUnit = "cycle",
  valueUnit = NULL
)
}
\arguments{
\item{data}{vector. A vector of any type representing the data of the series}

\item{start}{numeric. The beginning of the series in time, in cycle units.}

\item{end}{numeric. The end of the series in time, in cycle units.}

\item{duration}{numeric. Duration of the series in cycle units.}

\item{frequency}{numeric. Either frequency of period must be present to specify the
sampling rate of the signal. Frequency is espressed in number of observations
per cycle. While period is the duration of each observation in fractions of
cycle units. The period is equal to 1/frequency.}

\item{period}{numeric.}

\item{windowed}{list of three elements winSec, incSec, flex, describing eventual
windowing procedures used to create the series.}

\item{timeUnit}{character. A descriptor of the frequency cycle unit. Used only
in print functions.}

\item{valueUnit}{character. A descriptor of the unit of measurement of the series values. Used only
in print functions.}
}
\description{
given a STS starting are at 0000, if you have 5000ms of information
you only have samples for time 0, 1, 2, ... until 4999 so this would be represented as
00:00 - 00:04
Indeed time intervals are defined as half-open intervals
with start and end and window methods including the start value but not
the end value
}
\details{
Rational Time Series
rats is the creator for a nimble S3 class of regular time series.
rats are similar to \link[stats]{ts} or \link[zoo]{zooreg} classes, but these
classes have many quirks related to quarterly periods and other econometric
analytic traditions, while rats is more oriented towards physical signals.
The crucial difference is that for rats, time intervals are half-open intervals,
which means that they include the start value but not the end one. This way
rats(start=0, end=10, f=1) and rats(start=10, end=20, f=1) do not overlap.
Secondarily, rats are stored as a list with $x and $y values, storing respectively
the temporal information, and the values. These latter can be numeric or not.


* rats are legion: there is no singular form "rat" as there is no singular "timeserie"
* rats are small: rats is never capitalized.
* rats fit well together: rats(start=0, end=10, f=1) and rats(start=10, end=20, f=1) do
  not overlap and are combined to a . Time intervals are half-open intervals, including the start value
  but excluding the end one.
* rats are familiar: syntax is intuitive for ts() or zoo() users
* Holes fit in cheese, but not in rats: a rats' lenght must correspond to the rats'
  duration multiplied by the frequency. In other terms missing data in rats must
  be represented with NA
}
