% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_IO.R
\name{readCategories}
\alias{readCategories}
\title{Title}
\usage{
readCategories(dirPath, startCol, endCol, catName, namefilt = NA,
  removeSec = NULL, idOrder = c("dyadId", "session", "group"),
  idSep = "_", ...)
}
\arguments{
\item{dirPath}{}

\item{startCol}{}

\item{endCol}{}

\item{catName}{}

\item{namefilt}{}

\item{removeSec}{}

\item{idOrder}{}

\item{idSep}{}

\item{...}{}
}
\description{
Title
}
