% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_signals.R
\name{readDyadSignals}
\alias{readDyadSignals}
\title{Title}
\usage{
readDyadSignals(
  path,
  s1Col,
  s2Col,
  s1Name,
  s2Name,
  signalNames,
  SR,
  start,
  end,
  duration,
  unit,
  timeUnit = "seconds",
  pairBind = F,
  winTerpolate = list(winSec = NULL, incSec = NULL),
  namefilt = NA,
  idOrder = c(),
  idSep = "_",
  cores = TRUE,
  ...
)
}
\arguments{
\item{path}{}

\item{s1Col}{}

\item{s2Col}{}

\item{s1Name}{}

\item{s2Name}{}

\item{signalNames}{a vector of string defining the name of the signals (same order as s1Col, s2Col)}

\item{SR}{}

\item{start}{an optional vector of integers, or strings in mm:ss format which specify the time (in seconds) of the first observation of each file.
Note that by default time series will be set with start = 0, which is different from the default value of \code{\link[stats]{ts}}.}

\item{end}{an optional vector of integers, or strings in mm:ss format which specify the time (in seconds) of the last observation of each file.
If the data is longer or shorter, remove tail
 or pads with zeroes. Useful to cut signals when a session is over or to equalize lengths. The default value,
 FALSE, keeps the original length of the signal}

\item{duration}{an optional vector of integers, or strings in mm:ss format which specify the time (in seconds) of the duration of each file.
If the data is longer or shorter, remove tail or pads with zeroes. Only one between end or duration should be specified.}

\item{unit}{character. A descriptor of the unit of measurement of the series values.}

\item{timeUnit}{character. A descriptor of the time unit of the series values.
Defaults to seconds as in "samples per second"}

\item{pairBind}{if true, every two consequent files in path get matched to a single DyadSession, useful if the data of participant
1 and 2 are stored in separate files.}

\item{winTerpolate}{a list containing two numeric values, winSec and incSec. If the data to be read has been generated through a moving
window filter, it can be reversed to the given SR by setting the size (in seconds) of the window and its increment.}

\item{namefilt}{a string used to select only specific files in a given path}

\item{idOrder}{either NA or a character vector. idOrder is used to interpret the filenames to correctly identify and label the cases. 
If NA, no attempt to identify cases will be done. The character vector may contain one or more of the following strings in any given order: "id" (unique identifier of a dyad),
"session" (session number if the same dyad has multiple sessions),"group" (if the data must be devided across groups or experimental conditions),
"role" (if pairBind = TRUE, to assign the role of different files), or "x" (a placeholder to skip irrelevant information in filenames)
The strings can be abbreviated.}

\item{idSep}{character vector (or object which can be coerced to such) containing regular expression(s).
If idOrder is not NA, this will be used as separator to split the filenameas and identify "id", "session", and "group"
informations.}

\item{cores}{Either logical or integer. Sets how many cores should be used for the import. Values of 1, 0 or FALSE disable the multi-core code.}

\item{...}{additional options to be passed to read.table (es: skip, header, etc)}
}
\description{
Title
}
