% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{selectSignals}
\alias{selectSignals}
\title{Title
This function allows to subset only specific signals from an experiment}
\usage{
selectSignals(x, signals)
}
\arguments{
\item{x}{a DyadExperiment or DyadSession object}

\item{signals}{a vector of signal names. Es: c("SC", "HRV", "PACS")}
}
\description{
Title
This function allows to subset only specific signals from an experiment
}
