% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{setArg}
\alias{setArg}
\title{#' Window-based rescaling
#'
#' This function operates a rescale on a signal based on a rolling window.
#' It is useful to "auto zoom" a long signal
#'
#' @param x a ts or numeric object
#' @param winSec integer. The window size, in seconds
#' @param rangeMin lower boundary of rescaling
#' @param rangeMax upper boundary of rescaling
#'
#' @return
#' @export}
\usage{
setArg(arg, value, argList)
}
\arguments{
\item{arg}{}

\item{value}{}

\item{argList}{}
}
\description{

}
\details{
This is most useful when dealing with ... and setting default behaviours that 
can be overweritten by ...
}
