% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{setArg}
\alias{setArg}
\title{Set argument/attribute to a list without overwriting existing values}
\usage{
setArg(arg, value, argList)
}
\arguments{
\item{arg}{}

\item{value}{}

\item{argList}{}
}
\description{
This is most useful when dealing with ... and setting default behaviours that 
can be overweritten by ...
}
