% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{signalDecimate}
\alias{signalDecimate}
\title{signalDecimate
Takes an objects of class DyadSignal and downsamples it by taking a sample each oldSampRate / newSampRate}
\usage{
signalDecimate(signal, newSampRate)
}
\arguments{
\item{signal}{}

\item{newSampRate}{}
}
\description{
signalDecimate
Takes an objects of class DyadSignal and downsamples it by taking a sample each oldSampRate / newSampRate
}
