% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_filters.R
\name{signalFilter}
\alias{signalFilter}
\title{apply a function on each dyad's member signal}
\usage{
signalFilter(x, FUN, ..., newAttributes = NULL, signals = "all")
}
\arguments{
\item{FUN}{}

\item{...}{additional parameters of FUN}

\item{newAttributes}{named list of attributes to be added or replaced}

\item{signal}{DyadSignal Object}
}
\description{
apply a function on each dyad's member signal
}
\examples{

}
