% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_extract.R
\name{streamExtract}
\alias{streamExtract}
\title{Title}
\usage{
streamExtract(experiment, signal, sync = "none", streamKey = c("s1", "s2",
  "time", "sync", "lag", "zero"), FUN = mean, ...)
}
\arguments{
\item{experiment}{}

\item{sync}{}

\item{streamKey}{}

\item{FUN}{}

\item{...}{}

\item{by}{}
}
\description{
Title
}
