% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_classes.R
\name{[.DyadExperiment}
\alias{[.DyadExperiment}
\title{Extract or replace parts of a dyadexperiment}
\usage{
\method{[}{DyadExperiment}(x, i, name = NA)
}
\arguments{
\item{x}{}

\item{i}{}

\item{name}{The name of the new experiment}
}
\description{
Extract or replace parts of a dyadexperiment
}
