% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{timeMaster}
\alias{timeMaster}
\title{timeMaster
timeMaster allows to:
transform time from different formats to different formats.
add amounts of time to baseTime expressed in different formats.
This is verified. And useful. For reasons. :-D}
\usage{
timeMaster(
  baseTime,
  out = c("auto", "hour", "min", "sec"),
  add = 0,
  baseSep = "[\\\\.,:,;,\\\\,',\\",\\\\-]"
)
}
\arguments{
\item{baseTime}{}

\item{out}{}

\item{add}{}

\item{baseSep}{}
}
\description{
timeMaster
timeMaster allows to:
transform time from different formats to different formats.
add amounts of time to baseTime expressed in different formats.
This is verified. And useful. For reasons. :-D
}
