% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_classes.R
\name{window.DyadSignal}
\alias{window.DyadSignal}
\title{Time Windows}
\usage{
\method{window}{DyadSignal}(x, duration, ...)
}
\arguments{
\item{x}{}

\item{duration}{an alternative specification to end}

\item{...}{}
}
\value{

}
\description{
Time Windows
}
