% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP_utilities.R
\name{winInter}
\alias{winInter}
\title{Interpolate by window
Interpolate windowed data to original samplerate. 
useful to overlay computed indexes on original timeseries}
\usage{
winInter(windowsList, winSec, incSec, sampRate)
}
\arguments{
\item{winSec}{}

\item{incSec}{}

\item{sampRate}{}
}
\description{
Interpolate by window
Interpolate windowed data to original samplerate. 
useful to overlay computed indexes on original timeseries
}
